+++ pipelex/pipe_machinery/pipe_factory.py
@@ -5,7 +5,7 @@ from typing_extensions import runtime_checkable
 
 from pipelex.core.concepts.helpers import strip_markers_from_concept_ref_or_code
 from pipelex.core.concepts.native.concept_native import NativeConceptCode
-from pipelex.core.pipes.exceptions import PipeFactoryError, PipeFactoryErrorType
+from pipelex.core.pipes.exceptions import PipeFactoryError
 from pipelex.core.pipes.inputs.input_stuff_specs import InputStuffSpecs
 from pipelex.core.pipes.inputs.input_stuff_specs_factory import InputStuffSpecsFactory
 from pipelex.core.pipes.stuff_spec.exceptions import StuffSpecFactoryError
@@ -14,6 +14,7 @@ from pipelex.core.pipes.stuff_spec.stuff_spec_factory import StuffSpecFactory
 from pipelex.interpreter_hub import get_concept_library
 from pipelex.pipe_machinery.pipe_blueprint import PipeBlueprint, PipeCategory, PipeType
 from pipelex.runtime_hub import get_class_registry
+from pipelex.validation_error_types import PipeFactoryErrorType
 
 if TYPE_CHECKING:
     from pipelex.pipe_machinery.pipe_abstract import PipeAbstract