+++ tests/integration/pipelex/pipes/optionals/test_redundant_force_warning.py
@@ -8,7 +8,6 @@ from typing import Callable
 
 from pipelex.base_exceptions import ValidationErrorCategory
 from pipelex.core.memory.working_memory import WorkingMemory
-from pipelex.core.pipes.exceptions import PipeValidationErrorType
 from pipelex.core.stuffs.text_content import TextContent
 from pipelex.interpreter_hub import get_pipe_library
 from pipelex.pipe_controllers.sequence.pipe_sequence import PipeSequence
@@ -20,6 +19,7 @@ from pipelex.pipe_operators.func.pipe_func_blueprint import PipeFuncBlueprint
 from pipelex.pipeline.controller_taint import collect_controller_taint_analyses
 from pipelex.pipeline.optionality_warnings import build_optionality_warnings
 from pipelex.system.registries.func_registry import func_registry
+from pipelex.validation_error_types import PipeValidationErrorType
 
 _DOMAIN_CODE = "test_redundant_force"