+++ tests/unit/pipelex/mthds_parsing/test_handle_pipe_errors_suffix.py
@@ -1,5 +1,6 @@
-from pipelex.core.pipes.exceptions import PipeValidationError, PipeValidationErrorType
+from pipelex.core.pipes.exceptions import PipeValidationError
 from pipelex.mthds_parsing.handle_pipe_errors import categorize_pipe_validation_with_libraries_error
+from pipelex.validation_error_types import PipeValidationErrorType
 
 
 class TestRequiredProvidedSuffix: